works locally-20230402

+++ scripts/txt2img.py
@@ -61,7 +61,8 @@ def load_model_from_config(config, ckpt, verbose=False):
         print("unexpected keys:")
         print(u)
 
-    model.cuda()
+#    model.cuda()
+    model.cuda.half()
     model.eval()
     return model
